Add C4 diagram (#24)
# Pull Request

## Description

This pull request introduces a new script to generate C4 diagrams for
the project documentation. The script leverages the `diagrams` library
to create a visual representation of the system components and their
interactions.

Key changes include:

* Added a new script `scripts/diagram.py` to generate C4 diagrams using
the `diagrams` library. The script specifies Python version requirements
and dependencies.
* Defined system components such as `Person`, `Container`, and
`Database` within the script to represent different parts of the system.
* Created relationships between components to illustrate interactions,
such as a user interacting with a dashboard and a scanner creating a
data file.

Fixes #26