Init UV Project (#12)
# Pull Request

## Description

This pull request introduces a new Python project named `scanner` with a
basic structure and dependencies. The most important changes include the
addition of a `main.py` file with a simple `main` function and a
`pyproject.toml` file that defines the project metadata and
dependencies.

Project structure and initialization:

*
[`scanner/main.py`](diffhunk://#diff-dae1ba888af6415a083816435ab08d9c3ebf65718d657039d969bf973e028a37R1-R6):
Added a `main` function that prints "Hello from sloccount!" and an entry
point to run the `main` function if the script is executed directly.
*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586R1-R12):
Created a project configuration file with metadata, Python version
requirements, and dependencies on `gitpython` and `pygithub`.

Fixes #21