Add Scanner Git Hooks (#17)
# Pull Request

## Description

This pull request includes changes to the `lefthook.yml` and
`scanner/scanner.just` files to add new checks and update existing
commands for better organization and clarity.

### Enhancements to pre-commit checks:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR21-R24):
Added new sections for `Ruff Checks` and `Vulture Checks` to the
pre-commit configuration.

### Improvements to Ruff commands:

*
[`scanner/scanner.just`](diffhunk://#diff-3205dc7246237ab9f796a373b98d2390f8b03bc04ea76a8eb831ba83b55ee538L36-R42):
Updated the `ruff-fix` and `ruff-checks` commands to use the `scanner::`
namespace for better organization.

Fixes #44