Update CodeQL Actions (#57)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/code-checks.yml` file to upgrade the version of the
`github/codeql-action` used in various steps. The most important changes
include:

Version updates for `github/codeql-action`:

* Updated `github/codeql-action/upload-sarif` from `v3.28.10` to
`v3.28.11` in multiple steps to ensure compatibility and leverage any
improvements or fixes.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L113-R113)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L212-R212)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L255-R255)
* Updated `github/codeql-action/init` from `v3.28.10` to `v3.28.11` to
initialize CodeQL with the latest version.
* Updated `github/codeql-action/analyze` from `v3.28.10` to `v3.28.11`
to perform CodeQL analysis with the latest version.

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>