Add CodeLimit Job (#71)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow to
enforce code limits. The most important change is the addition of the
`run-code-limit` job to the `.github/workflows/code-checks.yml` file.

Changes in `.github/workflows/code-checks.yml`:

* Added a new job named `run-code-limit` to enforce code limits using
the `getcodelimit/codelimit-action` GitHub Action. This job includes
steps for checking out the repository and running the Code Limit action.

Fixes #43