Add Deploy Workflow (#75)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow for deploying
the project. The workflow is configured to run on pushes to the `main`
branch and on a daily schedule. It includes steps for setting up the
environment and running a scanner.

Key changes:

* Added a new workflow named "Deploy" in `.github/workflows/deploy.yml`
to automate deployment processes.
* Configured the workflow to trigger on pushes to the `main` branch and
on a daily schedule using cron.
* Set up permissions for reading contents and packages, and writing
statuses and security events.
* Included steps to checkout the repository, install Python and UV, set
up Just, install Python dependencies, and run the scanner.

Fixes #68