Analyse Repositories (#83)
# Pull Request

## Description

This pull request introduces the use of the `pygount` library to analyze
source code in cloned repositories and updates the project dependencies
accordingly. Below are the most important changes:

### Integration of `pygount` for Source Code Analysis:

*
[`scanner/app/__main__.py`](diffhunk://#diff-e370c8f7bf64354b82fe45c59db166c3b80722cbe29d17342bad06377b071639R4):
Added imports for `ProjectSummary` and `SourceAnalysis` from `pygount`.
Enhanced the `run_analyser` function to analyze files in cloned
repositories using `pygount`, logging the analysis results and creating
a project summary.
[[1]](diffhunk://#diff-e370c8f7bf64354b82fe45c59db166c3b80722cbe29d17342bad06377b071639R4)
[[2]](diffhunk://#diff-e370c8f7bf64354b82fe45c59db166c3b80722cbe29d17342bad06377b071639L35-R51)

### Dependency Updates:

*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586R9-R16):
Added `pygount==2.0.0` to the list of dependencies. Updated the version
of `ruff` in the optional development dependencies.
[[1]](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586R9-R16)
[[2]](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586R25-R27)

Fixes #6