Update SuperLinter EditorConfig Checker (#108)
# Pull Request

## Description

This pull request includes changes to the configuration files for the
super linter and the GitHub Actions workflow. The most important changes
include renaming a configuration file and adding a new environment
variable to the workflow.

Changes to configuration files:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L5-R5):
Renamed from `.ecrc` and corrected the key `SpacesAftertabs` to
`SpacesAfterTabs`.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R36):
Added the `EDITORCONFIG_FILE_NAME` environment variable to the `jobs:`
section.