Update Dependabot Schedule (#147)
# Pull Request

## Description

This pull request updates the Dependabot configuration to replace the
existing daily schedule with a cron-based schedule for various update
groups. The new schedule ensures updates are checked at a specific time
each day.

### Dependabot configuration updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L11-R12):
Changed the `schedule` for GitHub Actions dependencies to use a cron
expression (`30 7 * * *`) instead of specifying `daily`, `time`, and
`timezone`.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L29-R29):
Updated the `schedule` for Python dependencies in the `/diagrams` and
`/tests` directories to use the same cron expression.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L44-R43):
Modified the `schedule` for NPM dependencies in the `dashboard`
directory to adopt the cron-based scheduling.