Remove `.editorconfig-checker.json` (#171)
# Pull Request

## Description

This pull request removes the `.editorconfig-checker.json` configuration
file and updates the GitHub Actions workflow to reflect this removal.
These changes simplify the configuration by eliminating unused or
unnecessary settings.

### Configuration cleanup:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Deleted the file entirely, as it is no longer required for the project.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable from the
workflow configuration, as it referenced the deleted
`.editorconfig-checker.json` file.