docs: Update C4 Diagram to drawio file (#180)
# Pull Request

## Description

This pull request updates the `README.md` file to include a new section
on project architecture and a corresponding link in the table of
contents.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R9-R16):
Added an "Architecture" section with a C4 diagram image
(`docs/c4.drawio.png`) and updated the table of contents to include a
link to this new section.