ci: Upload Scanner Result (#185)
# Pull Request

## Description


This pull request updates the deployment workflow by adding a step to
upload an artifact. The change ensures that the output of the workflow
is saved as an artifact for later use.

* **Workflow enhancement**:
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R33-R37):
Added an "Upload Artifact" step using the `actions/upload-artifact`
action (v4.6.2) to upload `output.json` as an artifact named `output`.