ci: Add Scanner Output Scanning (#189)
# Pull Request

## Description

This pull request adds a new job to the deployment workflow to validate
the output of the scanner. The most important change is the introduction
of the `validate-scanner-output` job, which ensures that the scanner's
output adheres to the expected schema.

### Workflow Enhancements:
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R41-R60):
Added a new job, `validate-scanner-output`, to validate the scanner's
output. This job runs after `run-scanner` and includes steps to check
out the repository, set up test dependencies, download the scanner's
output artifact, and validate its schema using the `just
tests::validate-schema` command.