docs: Add Project Status (#192)
# Pull Request

## Description

This pull request introduces a new validation step in the deployment
workflow and updates the `README.md` with a badge indicating the
project's development status.

### Workflow Enhancements:
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R41-R60):
Added a new job, `validate-scanner-output`, to validate the scanner's
output. This includes steps for checking out the repository, setting up
test dependencies, downloading artifacts, and running a schema
validation command.

### Documentation Updates:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3-R4):
Added a badge to display the project's active development status.