Force colour on workflow (#226)
# Pull Request

## Description

This pull request includes updates to configuration files and dependency
versions across multiple components. The key changes involve setting
environment variables for GitHub workflows and updating the required
version of a dependency (`uv`) in two separate files.

### Workflow configuration updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R15-R17):
Added an environment variable `FORCE_COLOR` with a value of `1` to
enforce colored output during code checks.

### Dependency version updates:
*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586L16-R16):
Updated the `required-version` of `uv` from `0.7.8` to `0.7.13` to
ensure compatibility with newer features or fixes.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL8-R8):
Similarly updated the `required-version` of `uv` from `0.7.8` to
`0.7.13` in the test configuration.