Upgrade Python Versions (#227)
# Pull Request

## Description

This pull request updates dependencies in the `pyproject.toml` files for
the `scanner` project and its tests. The changes primarily involve
upgrading existing dependencies and adding a new development dependency
(`ty`).

Dependency updates:

*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586L8-R19):
Upgraded `structlog` to version `25.4.0`, `pytest` to version `8.4.0`,
`pytest-cov` to version `6.2.1`, and `ruff` to version `0.12.0`. Added
`ty==0.0.1a11` as a new development dependency.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL5-R10):
Upgraded `pytest` to version `8.4.0` and `ruff` to version `0.12.0`.
Added `ty==0.0.1a11` as a new dependency.